requirements.txt
bcrypt==3.2.2
cryptography==44.0.2
fastapi==0.111.0
fastapi-cli==0.0.4
passlib==1.7.4
pydantic==2.7.4
pydantic-settings==2.7.1
pydantic_core==2.18.4
pyotp==2.9.0
python-jose==3.4.0
qrcode==8.2
uvicorn==0.30.1